delete working but cant get the styling right
