+++ src/components/profile/venuemanger/YourVenues.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import { useParams, Link } from 'react-router-dom';
-import { PROFILE_URL } from '../../../constant/api';
+import { PROFILE_URL, ALLVENUES_URL } from '../../../constant/api';
 import { useUser } from '../../../hooks/type/UserContext';
 import { useFetch } from '../../../hooks/useFetch';
 import { format } from 'date-fns';
 import DefaultImage from '../../../assets/Default.png';
+import DeleteButton from '../../delete/Delete';
 
 function YourVenues() {
   const { name: profileName } = useParams();
@@ -67,12 +68,26 @@ function YourVenues() {
               >
                 View Venue
               </Link>
-              <Link
-                to={`/edityourvenue/${venue.id}`} 
-                className="block bg-dark-blue text-center text-white p-2 rounded-md hover:bg-white hover:text-dark-blue border border-dark-blue"
-              >
-               Edit Venue
-              </Link>
+              <div className='flex'>
+                <div className="flex-1">
+                  <Link
+                    to={`/edityourvenue/${venue.id}`} 
+                    className="block bg-dark-blue text-center text-white m-5 p-2 rounded-md hover:bg-white hover:text-dark-blue border border-dark-blue "
+                  >
+                    Edit Venue
+                  </Link>
+                </div>
+                <div className="flex-1">
+                  <DeleteButton
+                    onDelete={() => {}}
+                    user={user}
+                    apiEndpoint={`${ALLVENUES_URL}/${venue.id}`}  
+                    invalidationKey={['data']}
+                    className="flex p-2 w-full border"
+                  />
+                </div>
+              </div>
+
             </div>
 
           </div>
